+++ packages/webr/src/mount.c
@@ -15,6 +15,14 @@
     Rf_error("`" #arg "` can't be `NA`.");               \
   }
 
+#define CHECK_LOGICAL(arg) \
+  if (!Rf_isLogical(arg) || LENGTH(arg) != 1) {          \
+    Rf_error("`" #arg "` must be a logical.");           \
+  }                                                      \
+  if (LOGICAL(arg)[0] == NA_LOGICAL){                    \
+    Rf_error("`" #arg "` can't be `NA`.");               \
+  }
+
 SEXP ffi_mount_workerfs(SEXP source, SEXP mountpoint) {
 #ifdef __EMSCRIPTEN__
   CHECK_STRING(source);
@@ -68,6 +76,47 @@ SEXP ffi_mount_nodefs(SEXP source, SEXP mountpoint) {
     }
   }, R_CHAR(STRING_ELT(source, 0)), R_CHAR(STRING_ELT(mountpoint, 0)));
 
+  return R_NilValue;
+#else
+  Rf_error("Function must be running under Emscripten.");
+#endif
+}
+
+SEXP ffi_mount_idbfs(SEXP mountpoint) {
+#ifdef __EMSCRIPTEN__
+  CHECK_STRING(mountpoint);
+
+  EM_ASM({
+    // Stop if we're not able to use a IDBFS filesystem object
+    if (typeof IN_NODE === 'boolean' && IN_NODE === true) {
+      const msg = Module.allocateUTF8OnStack(
+        'The `IDBFS` filesystem object can only be used when running in a web browser.'
+      );
+      Module._Rf_error(msg);
+    }
+    const mountpoint = UTF8ToString($0);
+    try {
+      Module.FS.mount(Module.FS.filesystems.IDBFS, {}, mountpoint);
+    } catch (e) {
+      let msg = e.message;
+      if (e.name === "ErrnoError" && e.errno === 10) {
+        const dir = Module.UTF8ToString($0);
+        msg = "Unable to mount directory, `" + dir + "` is already mounted.";
+      }
+      Module._Rf_error(Module.allocateUTF8OnStack(msg));
+    }
+  }, R_CHAR(STRING_ELT(mountpoint, 0)));
+
+  return R_NilValue;
+#else
+  Rf_error("Function must be running under Emscripten.");
+#endif
+}
+
+SEXP ffi_syncfs(SEXP populate) {
+#ifdef __EMSCRIPTEN__
+  CHECK_LOGICAL(populate);
+  EM_ASM({ Module.FS.syncfs($0, () => {}) }, LOGICAL(populate)[0]);
   return R_NilValue;
 #else
   Rf_error("Function must be running under Emscripten.");