+++ src/esbuild.ts
@@ -45,7 +45,7 @@ function build(input: string, output: string, platform: esbuild.Platform, minify
 
 const outputs = {
   browser: [
-    build('repl/App.tsx', '../dist/repl.mjs', 'neutral', prod),
+    build('repl/App.tsx', '../dist/repl.mjs', 'browser', prod),
     build('webR/chan/serviceworker.ts', '../dist/webr-serviceworker.js', 'browser', false),
     build('webR/webr-worker.ts', '../dist/webr-worker.js', 'node', false),
     build('webR/webr-main.ts', '../dist/webr.mjs', 'neutral', prod),
@@ -86,7 +86,7 @@ if (serve) {
                   res.writeHead(proxyRes.statusCode!, {
                     ...proxyRes.headers,
                     'cross-origin-opener-policy': 'same-origin',
-                    'cross-origin-embedder-policy': 'require-corp',
+                    'cross-origin-embedder-policy': 'credentialless',
                     'cross-origin-resource-policy': 'cross-origin',
                   });
                   proxyRes.pipe(res, { end: true });