+++ src/repl/components/Editor.tsx
@@ -6,24 +6,44 @@ import { EditorState, Compartment, Prec } from '@codemirror/state';
 import { autocompletion, CompletionContext } from '@codemirror/autocomplete';
 import { keymap } from '@codemirror/view';
 import { indentWithTab } from '@codemirror/commands';
+import { Panel } from 'react-resizable-panels';
 import { FilesInterface, TerminalInterface } from '../App';
 import { r } from 'codemirror-lang-r';
-import './Editor.css';
-import { WebRDataJsAtomic } from '../../webR/robj';
+import { NamedObject, WebRDataJsAtomic } from '../../webR/robj';
+import DataGrid from 'react-data-grid';
 import * as utils from './utils';
+import 'react-data-grid/lib/styles.css';
+import './Editor.css';
 
 const language = new Compartment();
 const tabSize = new Compartment();
 
-export type EditorFile = {
-  name: string;
-  path: string;
-  ref: {
-    editorState: EditorState;
-    scrollTop?: number;
-    scrollLeft?: number;
+type EditorBase = { name: string, readOnly: boolean };
+type EditorData = EditorBase & {
+  type: "data",
+  data: {
+    columns: { key: string, name: string }[];
+    rows: { [key: string]: string }[];
   }
 };
+
+type EditorHtml = EditorBase & {
+  path: string;
+  type: "html",
+  readOnly: boolean,
+  frame: HTMLIFrameElement,
+};
+
+type EditorFile = EditorBase & {
+  path: string;
+  type: "text",
+  readOnly: boolean,
+  editorState: EditorState;
+  scrollTop?: number;
+  scrollLeft?: number;
+};
+
+export type EditorItem = EditorData | EditorHtml | EditorFile;
 
 export function FileTabs({
   files,
@@ -31,7 +51,7 @@ export function FileTabs({
   setActiveFileIdx,
   closeFile
 }: {
-  files: EditorFile[];
+  files: EditorItem[];
   activeFileIdx: number;
   setActiveFileIdx: React.Dispatch<React.SetStateAction<number>>;
   closeFile: (e: React.SyntheticEvent, index: number) => void;
@@ -66,7 +86,7 @@ export function FileTabs({
             className="editor-close"
             aria-label={`Close ${f.name}`}
             onClick={(e) => {
-              if (!f.ref.editorState.readOnly && !confirm('Close ' + f.name + '?')) {
+              if (!f.readOnly && !confirm('Close ' + f.name + '?')) {
                 e.stopPropagation();
                 return;
               }
@@ -91,16 +111,19 @@ export function Editor({
   filesInterface: FilesInterface;
 }) {
   const editorRef = React.useRef<HTMLDivElement | null>(null);
+  const htmlRef = React.useRef<HTMLDivElement | null>(null);
   const [editorView, setEditorView] = React.useState<EditorView>();
-  const [files, setFiles] = React.useState<EditorFile[]>([]);
+  const [files, setFiles] = React.useState<EditorItem[]>([]);
   const [activeFileIdx, setActiveFileIdx] = React.useState(0);
   const runSelectedCode = React.useRef((): void => {
     throw new Error('Unable to run code, webR not initialised.');
   });
 
   const activeFile = files[activeFileIdx];
-  const isRFile = activeFile && activeFile.name.endsWith('.R');
-  const isReadOnly = activeFile && activeFile.ref.editorState.readOnly;
+  const isScript = activeFile && activeFile.type === "text" && activeFile.path.endsWith('.R');
+  const isData = activeFile && activeFile.type === "data";
+  const isHtml = activeFile && activeFile.type === "html";
+  const isReadOnly = activeFile && activeFile.readOnly;
 
   const completionMethods = React.useRef<null | {
     assignLineBuffer: RFunction;
@@ -141,7 +164,7 @@ export function Editor({
     if (from === to && !context.explicit) {
       return null;
     }
-    await completionMethods.current.assignLineBuffer(line);
+    await completionMethods.current.assignLineBuffer(line.replace(/\)+$/, ""));
     await completionMethods.current.assignToken(text);
     await completionMethods.current.assignStart(from + 1);
     await completionMethods.current.assignEnd(to + 1);
@@ -151,7 +174,7 @@ export function Editor({
       if (!val) {
         throw new Error('Missing values in completion result.');
       }
-      return { label: val };
+      return { label: val, boost: val.endsWith("=") ? 10 : 0 };
     });
 
     return { from: from, options };
@@ -179,11 +202,18 @@ export function Editor({
 
   const closeFile = (e: React.SyntheticEvent, index: number) => {
     e.stopPropagation();
+    const item = files[index];
+    if (item.type === "html") {
+      item.frame.remove();
+    }
+
     const updatedFiles = [...files];
     updatedFiles.splice(index, 1);
     setFiles(updatedFiles);
-    const prevFile = activeFileIdx - 1;
-    setActiveFileIdx(prevFile < 0 ? 0 : prevFile);
+    if (index <= activeFileIdx) {
+      const prevFile = activeFileIdx - 1;
+      setActiveFileIdx(prevFile < 0 ? 0 : prevFile);
+    }
   };
 
   React.useEffect(() => {
@@ -211,9 +241,11 @@ export function Editor({
     if (!editorView || !activeFile) {
       return;
     }
-    activeFile.ref.editorState = editorView.state;
-    activeFile.ref.scrollTop = editorView.scrollDOM.scrollTop;
-    activeFile.ref.scrollLeft = editorView.scrollDOM.scrollLeft;
+    if (activeFile.type === "text") {
+      activeFile.editorState = editorView.state;
+      activeFile.scrollTop = editorView.scrollDOM.scrollTop;
+      activeFile.scrollLeft = editorView.scrollDOM.scrollLeft;
+    }
   }, [activeFile, editorView]);
 
   const runFile = React.useCallback(() => {
@@ -234,7 +266,7 @@ export function Editor({
   }, [syncActiveFileState, editorView]);
 
   const saveFile: React.MouseEventHandler<HTMLButtonElement> = React.useCallback(() => {
-    if (!editorView) {
+    if (!editorView || activeFile.type !== "text") {
       return;
     }
 
@@ -264,9 +296,9 @@ export function Editor({
     setFiles([{
       name: 'Untitled1.R',
       path: '/home/web_user/Untitled1.R',
-      ref: {
-        editorState: state,
-      }
+      type: 'text',
+      readOnly: false,
+      editorState: state,
     }]);
 
     return function cleanup() {
@@ -279,9 +311,59 @@ export function Editor({
    * opening files they are displayed in this codemirror instance.
    */
   React.useEffect(() => {
+    filesInterface.openDataInEditor = (title: string, data: NamedObject<WebRDataJsAtomic<string>>) => {
+      // If data is there switch to that tab instead
+      const existsIndex = files.findIndex((f) => f.name === title);
+      if (existsIndex >= 0) {
+        setActiveFileIdx(existsIndex);
+        return;
+      }
+
+      syncActiveFileState();
+
+      const columns = Object.keys(data).map((key) => {
+        return {key, name: key === "row.names" ? "" : key};
+      });
+
+      const rows = Object.entries(data).reduce((a, entry) => {
+        entry[1].values.forEach((v, j) => a[j] = Object.assign(a[j] || {}, { [entry[0]!]: v }));
+        return a;
+      }, []);
+
+      const updatedFiles = [...files];
+      const index = updatedFiles.push({
+        name: title,
+        type: "data",
+        readOnly: true,
+        data: { columns, rows }
+      });
+      setFiles(updatedFiles);
+      setActiveFileIdx(index - 1);
+    };
+
+    filesInterface.openHtmlInEditor = (src: string, path: string) => {
+      syncActiveFileState();
+
+      const frame = document.createElement('iframe');
+      frame.srcdoc = src;
+      frame.className = "html-viewer";
+      htmlRef.current!.appendChild(frame);
+
+      const updatedFiles = [...files];
+      const index = updatedFiles.push({
+        name: 'Viewer',
+        path,
+        type: "html",
+        readOnly: true,
+        frame,
+      });
+      setFiles(updatedFiles);
+      setActiveFileIdx(index - 1);
+    };
+
     filesInterface.openFileInEditor = (name: string, path: string, readOnly: boolean) => {
       // Don't reopen the file if it's already open, switch to that tab instead
-      const existsIndex = files.findIndex((f) => f.path === path);
+      const existsIndex = files.findIndex((f) => "path" in f && f.path === path);
       if (existsIndex >= 0) {
         setActiveFileIdx(existsIndex);
         return Promise.resolve();
@@ -303,12 +385,12 @@ export function Editor({
         const index = updatedFiles.push({
           name,
           path,
-          ref: {
-            editorState: EditorState.create({
-              doc: content,
-              extensions,
-            }),
-          }
+          type: "text",
+          readOnly,
+          editorState: EditorState.create({
+            doc: content,
+            extensions,
+          }),
         });
         setFiles(updatedFiles);
         setActiveFileIdx(index - 1);
@@ -317,18 +399,32 @@ export function Editor({
   }, [files, filesInterface]);
 
   React.useEffect(() => {
+    if (activeFile && activeFile.type === "html") {
+      activeFile.frame.classList.remove("d-none");
+    }
+    // Before switching activeFile, hide this HTML
+    return function cleanup() {
+      if (activeFile && activeFile.type === "html") {
+        activeFile.frame.classList.add("d-none");
+      }
+    };
+  }, [activeFile]);
+
+  React.useEffect(() => {
     if (!editorView || files.length === 0) {
       return;
     }
     // Update the editor's state and scroll position for currently active file
-    editorView.setState(activeFile.ref.editorState);
-    editorView.requestMeasure({
-      read: () => {
-        editorView.scrollDOM.scrollTop = activeFile.ref.scrollTop ?? 0;
-        editorView.scrollDOM.scrollLeft = activeFile.ref.scrollLeft ?? 0;
-        return editorView.domAtPos(0).node;
-      }
-    });
+    if (activeFile.type === "text") {
+      editorView.setState(activeFile.editorState);
+      editorView.requestMeasure({
+        read: () => {
+          editorView.scrollDOM.scrollTop = activeFile.scrollTop ?? 0;
+          editorView.scrollDOM.scrollLeft = activeFile.scrollLeft ?? 0;
+          return editorView.domAtPos(0).node;
+        }
+      });
+    }
 
     // Update accessibility labelling
     const container = editorView.contentDOM.parentElement;
@@ -341,13 +437,14 @@ export function Editor({
     };
   }, [files, syncActiveFileState, activeFile, editorView]);
 
-  const displayStyle = files.length === 0 ? { display: 'none' } : undefined;
-
   return (
-    <div role="region"
+    <Panel
+      id="editor"
+      role="region"
       aria-label="Editor Pane"
-      className="editor"
-      style={displayStyle}
+      order={1}
+      minSize={20}
+      className={files.length === 0 ? "d-none" : ""}
     >
       <div className="editor-header">
         <FileTabs
@@ -360,29 +457,45 @@ export function Editor({
       <div
         aria-label="Editor"
         aria-describedby="editor-desc"
-        className="editor-container"
+        className={`editor-container ${(isData || isHtml) ? "d-none" : ""}`}
         ref={editorRef}
       >
       </div>
-      <p style={{ display: 'none' }} id="editor-desc">
+      <p className="d-none" id="editor-desc">
         This component is an instance of the <a href="https://codemirror.net/">CodeMirror</a> interactive text editor.
         The editor has been configured so that the Tab key controls the indentation of code.
         To move focus away from the editor, press the Escape key, and then press the Tab key directly after it.
         Escape and then Shift-Tab can also be used to move focus backwards.
       </p>
+      {(isData && activeFile.data) &&
+        <DataGrid
+          aria-label="Data viewer"
+          columns={activeFile.data.columns}
+          rows={activeFile.data.rows}
+          className="data-container"
+          defaultColumnOptions={{
+            sortable: true,
+            resizable: true
+          }}
+        />
+      }
+      <div
+        className={`html-viewer-container ${isHtml ? "" : "d-none"}`}
+        ref={htmlRef}
+      ></div>
       <div
         role="toolbar"
         aria-label="Editor Toolbar"
         className="editor-actions"
       >
-        {isRFile && <button onClick={runFile}>
+        {isScript && <button onClick={runFile}>
           <FaPlay aria-hidden="true" className="icon" /> Run
         </button>}
         {!isReadOnly && <button onClick={saveFile}>
           <FaRegSave aria-hidden="true" className="icon" /> Save
         </button>}
       </div>
-    </div>
+    </Panel>
   );
 }