+++ src/repl/components/Plot.tsx
@@ -2,15 +2,21 @@ import React from 'react';
 import './Plot.css';
 import { PlotInterface } from '../App';
 import { FaArrowCircleLeft, FaArrowCircleRight, FaRegSave, FaTrashAlt } from 'react-icons/fa';
+import { Panel, ImperativePanelHandle } from 'react-resizable-panels';
+import { WebR } from '../../webR/webr-main';
 
 export function Plot({
+  webR,
   plotInterface,
 }: {
+  webR: WebR;
   plotInterface: PlotInterface;
 }) {
   const plotContainerRef = React.useRef<HTMLDivElement | null>(null);
+  const panelRef = React.useRef<ImperativePanelHandle | null>(null);
   const canvasRef = React.useRef<HTMLCanvasElement | null>(null);
   const canvasElements = React.useRef<HTMLCanvasElement[]>([]);
+  const plotSize = React.useRef<{width: number, height: number}>({width: 1008, height: 1008});
   const [selectedCanvas, setSelectedCanvas] = React.useState<number | null>(null);
 
   // Register the current canvas with the plotting interface so that when the
@@ -27,14 +33,42 @@ export function Plot({
     plotInterface.newPlot = () => {
       const plotNumber = canvasElements.current.length + 1;
       const canvas = document.createElement('canvas');
-      canvas.setAttribute('width', '1008');
-      canvas.setAttribute('height', '1008');
+      canvas.setAttribute('width', String(plotSize.current.width * 2));
+      canvas.setAttribute('height', String(plotSize.current.height * 2));
       canvas.setAttribute('aria-label', `R Plot ${plotNumber}`);
       canvasRef.current = canvas;
       canvasElements.current.push(canvas);
       setSelectedCanvas(plotNumber - 1);
     };
+
+    // Resize the canvas() device when the plotting pane changes size
+    plotInterface.resize = (direction, px) => {
+      plotSize.current[direction] = px;
+      void webR.init().then(async () => {
+        await webR.evalRVoid(`
+          # Close any active canvas devices
+          repeat {
+            devices <- dev.list()
+            idx <- which(names(devices) == "canvas")
+            if (length(idx) == 0) {
+              break
+            }
+            dev.off(devices[idx[1]])
+          }
+          # Set canvas size for future devices
+          options(webr.fig.width = ${plotSize.current.width}, webr.fig.height = ${plotSize.current.height})
+      `, { env: {} });
+      });
+    };
   }, [plotInterface]);
+
+  const onResize = (size:number) => plotInterface.resize("height", size * window.innerHeight / 100);
+  React.useEffect(() => {
+    window.addEventListener("resize", () => {
+      if (!panelRef.current) return;
+      onResize(panelRef.current.getSize());
+    });
+  }, []);
 
   // Update the plot container to display the currently selected canvas element
   React.useEffect(() => {
@@ -46,6 +80,7 @@ export function Plot({
     } else {
       const canvas = canvasElements.current[selectedCanvas];
       plotContainerRef.current.replaceChildren(canvas);
+      plotContainerRef.current.style.aspectRatio = `${canvas.width} / ${canvas.height}`;
     }
   }, [selectedCanvas]);
 
@@ -67,7 +102,14 @@ export function Plot({
   const prevPlot = () => setSelectedCanvas((selectedCanvas === null) ? null : selectedCanvas - 1);
 
   return (
-    <div role="region" aria-label="Plotting Pane" className="plot">
+    <Panel
+      id="plot"
+      role="region"
+      aria-label="Plotting Pane"
+      minSize={20}
+      onResize={onResize}
+      ref={panelRef}
+    >
       <div className="plot-header">
         <div role="toolbar" aria-label="Plotting Toolbar" className="plot-actions">
           <button
@@ -111,7 +153,7 @@ export function Plot({
       <div className='plot-background'>
         <div ref={plotContainerRef} className="plot-container"></div>
       </div>
-    </div>
+    </Panel>
   );
 }