+++ src/webR/webr-chan.ts
@@ -5,7 +5,7 @@ import { Message } from './chan/message';
 import { UUID as ShelterID } from './chan/task-common';
 import { EmPtr } from './emscripten';
 import { WebRPayloadWorker, WebRPayloadPtr } from './payload';
-import { RType, RCtor, WebRData } from './robj';
+import { RType, RCtor, WebRData, WebRDataJsAtomic } from './robj';
 import type { FSType, FSMountOptions } from './webr-main';
 
 export { isUUID as isShelterID, UUID as ShelterID } from './chan/task-common';
@@ -26,10 +26,10 @@ export interface CallRObjectMethodMessage extends Message {
  */
 export interface InstallPackagesOptions {
   /**
-   * The R package repository from which to download packages.
+   * The R package repositories from which to download packages.
    * Default: The configured default webR package repository.
    */
-  repos?: string;
+  repos?: string | string[];
   /**
    * If `true`, do not output downloading messages.
    * Default: `false`.
@@ -44,9 +44,9 @@ export interface InstallPackagesOptions {
 
 /** @internal */
 export interface InstallPackagesMessage extends Message {
-  type: 'installPackage';
+  type: 'installPackages';
   data: {
-    name: string;
+    name: string | string[];
     options: InstallPackagesOptions;
   };
 }
@@ -157,6 +157,12 @@ export interface FSMountMessage extends Message {
 }
 
 /** @internal */
+export interface FSSyncfsMessage extends Message {
+  type: 'syncfs';
+  data: { populate: boolean };
+}
+
+/** @internal */
 export interface FSReadFileMessage extends Message {
   type: 'readFile';
   data: {
@@ -228,4 +234,19 @@ export interface PagerMessage extends Message {
     title: string;
     deleteFile: boolean;
   };
+}
+
+export interface ViewMessage extends Message {
+  type: 'view';
+  data: {
+    data: {
+      [key: string]: WebRDataJsAtomic<string>;
+    };
+    title: string;
+  };
+}
+
+export interface BrowseMessage extends Message {
+  type: 'browse';
+  data: { url: string };
 }